+++ src/SFA.DAS.LevyTransferMatching/Data/ValueObjects/CreateApplicationProperties.cs
@@ -9,7 +9,7 @@ namespace SFA.DAS.LevyTransferMatching.Data.ValueObjects
         public string Details { get; set; }
         public string StandardId { get; set; }
         public string StandardTitle { get; set; }
-        public byte StandardLevel { get; set; }
+        public int StandardLevel { get; set; }
         public int StandardDuration { get; set; }
         public int StandardMaxFunding { get; set; }
         public string StandardRoute { get; set; }
